blood_analysis/pyproject.toml
[project]
name = "blood-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.34.1",
    "python-dotenv>=1.0.1",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.5",
    "streamlit>=1.37.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
